requirements.txt
fastapi
uvicorn
celery
redis
pandas
fyers-apiv3
python-dotenv
statsmodels